This MR integrates Sentry into `message_logger`.
As `message_logger` is shared open source on GitHub we cannot introduce a hard-dependency so the Sentry dependency is disabled by default in case the dependencies `sentry_native` or `anymal_sentry_native` are not found.

Also, the feature toggle `error_reporting` must contain `automatic` in order for Sentry to be enabled.

How Sentry is integrated:

* Error and fatal logs trigger an error/fatal event.
* Debug, info, and warning logs are treated as breadcrumbs which are attached in case there is a subsequent error/fatal event. The current threshold is 100 breadcrumbs. We'll likely need to further fine tune the integration in the future.
* Of course, a programmer can still emit Sentry events without having to go through MELO. The integration here is just a "one catches all" trial to immediately get a high coverage of errors.
* Adding attachments will follow, that will require us to configure the individual processes.

**Related Issues** Closes #

I needed to remove the dependency of `anymal_software_metadata` on `watchdog` to avoid a circular dependency: `anymal_software_metadata` -> `watchdog` -> `param_io` -> `message_logger` -> `anymal_software_metadata`. This should not be an issue as I don't expect the `anymal_software_metadata` unit test to deadlock.

The changes are tested in simulation (issues are found [here](https://anybotics.sentry.io/projects/native-devel/?issuesType=all&project=4505687018831872), as I was working on a devel branch != nightly / release they were pushed to the `native-devel` project to avoid spamming the `native` project with noise).

Once merged I'll test this on a robot. This will be much easier because I can just enable the feature toggle instead of having to compile the full stack with `install` and then copy it to the robot using `anymal_docker`...

<!-- Please make sure to complete the following checklist before starting the review process.
     Use the following icons:
     YES: [x]
     NO:  [ ] -->
- [ ] Automated test in place :rocket:
- [ ] Repeated manual test in place (shakeout, release tests) :scroll: :nerd:
- One-time manual test performed :one: :scream:
  - [x] in simulation :computer:
  - [ ] on robot :robot:
  - [ ] replaying a dataset :vhs:
- [x] I performed the following steps if applicable:
  - [ ] Update the version of any package with modified code according to the [software development guidelines](https://anybotics.docs.anymal.com/user_manual/anymal_c100/master/documents/software_development_guide/main.html#semantic-versioning)
  - [ ] Create/update documentation (incl. [release notes](https://git.anybotics.com/anybotics/anybotics/-/tree/master/documentation/source/documents/release_notes), [user manual changelog](https://git.anybotics.com/anybotics/anybotics/-/blob/master/documentation/source/documents/anymal_user_guide/information.inc)).
  - [ ] Create/update tests and test instructions ([manual release tests](https://www.notion.so/anybotics/Manual-Testing-7ef73c0e78b24c6aa7bddabbfb73b636?pvs=4#a10727f20599466684d9caa57abb890c), [shakeout](https://git.anybotics.com/anybotics/anybotics/-/blob/master/.gitlab/issue_templates/Shakeout%20Test.md))
  - [ ] Update potential downstream dependencies, e.g. QAP
  - [ ] Label related issues with ~"W::In Review".
  - [x] Conform by the [style guide](https://anybotics.github.io/styleguide/cppguide.html).
  - [x] Conform by the [software development guidelines](https://anymal-research.docs.anymal.com/user_manual/anymal_c100/documents/software_development_guide/main.html#development-guidelines).

GitOrigin-RevId: f14a9e2572ee594148537aefc27debf569778e82